+++ src/command_modules/azure-cli-consumption/azure/cli/command_modules/consumption/_validators.py
@@ -87,5 +87,3 @@ def validate_budget_parameters(namespace):
 
     if namespace.amount < 0:
         raise CLIError("usage error: --parameters.time_period.amount must be greater than 0")
-
-