+++ src/command_modules/azure-cli-consumption/azure/cli/command_modules/consumption/_help.py
@@ -53,3 +53,42 @@ helps['consumption pricesheet show'] = """
     type: command
     short-summary: Show the price sheet for Azure subscription within a current billing period or specified billing period
 """
+
+helps['consumption marketplace'] = """
+    type: group
+    short-summary: Inspect the marketplace by Azure subscription within current or specified billing period.
+"""
+
+helps['consumption marketplace list'] = """
+    type: command
+    short-summary: Show the marketplace for Azure subscription within a current billing period or specified billing period
+"""
+
+helps['consumption budget'] = """
+    type: group
+    short-summary: Offers ability to create, get, update, and delete budgets for Azure subscription optionally by resource group name
+"""
+
+helps['consumption budget list'] = """
+    type: command
+    short-summary: Offers ability to create, get, update, and delete budgets for Azure subscription optionally by resource group name
+"""
+
+helps['consumption budget show'] = """
+    type: command
+    short-summary: Show budget for Azure subscription by budget name and, optionally by, resource group name
+"""
+
+helps['consumption budget create'] = """
+    type: command
+    short-summary: Create budget for Azure subscription and associate with resource group name
+"""
+
+helps['consumption budget update'] = """
+    type: command
+    short-summary: Update budget for Azure subscription and associate with resource group name
+"""
+helps['consumption budget delete'] = """
+    type: command
+    short-summary: Delete budget for Azure subscription optionally by resource group name
+"""