+++ src/command_modules/azure-cli-consumption/HISTORY.rst
@@ -2,6 +2,10 @@
 
 Release History
 ===============
+0.2.3
++++++
+* Added new commands for marketplace and budget API. Update to pricesheet to include top filter
+
 0.2.2
 +++++
 * Added new commands for pricesheet API. Also, updated the existing Usage Details and Reservation Details API request format.